Add files via upload
This work is on data classification techniques in developing a model to detect data intrusion in a network. Work file is a sample extracted from data files used for The Third International Knowledge Discovery and Data Mining Tools Competition, kddcupData. The sample file is a csv file and SPSS modeler 18.3 is used for the data analysis to classify the attacks. Analysis was done on 98327 records with 42 attributes. Of these huge data, concentration was on the attribute ‘connection_type’ in which 22 type of attacks were present in the data set. Out of this, safe is the one with connection type ‘normal’ and remaining 21 are the records with unsafe attacks. With the given data, a model has been trained and developed for classifying any incoming data into the different attacks