Add files via upload
Diabetes is a serios chronic disease which is increasing rapidly now a days. This disease is increasing its strength and moving upon every individual who is not disciplined to one’s health habits, no leaving even the children. In this project, an analysis of this disease is been made by imposing few data mining tasks on a chosen dataset. This dataset is collection of responses from the individuals on the status of diabetes based on few indices like BMI, cholesterol, age, etc. This project work is an attempt in which relevant data mining tasks such as prediction, association and clustering are imposed on the dataset to evaluate the presence of diabetes, association rules between the factors and clustering into few groups with similar data attributes. The resultant data is used for further analysis on evaluating and deciding which factors are more prevailing, the dependents and the common groups.